+++ js/form.js
@@ -1,13 +1,7 @@
 import { resetScale } from './scale.js';
 import { isEscapeKey } from './util.js';
-
-const imgEditorForm = document.querySelector('.img-upload__form');
-const imgEditorElement = document.querySelector('.img-upload__overlay');
-const imgEditorOpenElement = document.querySelector('#upload-file');
-const imgEditorcloseElement = imgEditorElement.querySelector('#upload-cancel');
-const bodyElement = document.querySelector('body');
-const commentElement = imgEditorForm.querySelector('.text__description');
-const EMPTY_VALUE = '';
+import { imgEditorForm, imgEditorElement, imgEditorOpenElement, imgEditorcloseElement, bodyElement, commentElement } from './dom_elements.js';
+import { EMPTY_VALUE } from './data.js';
 
 const ImgUpload = () => {
   const onEditorEscKeydown = (evt) => {
@@ -79,4 +73,4 @@ const validateImgEditorForm = () => {
   });
 };
 
-export {ImgUpload, validateImgEditorForm, imgEditorForm};
+export { ImgUpload, validateImgEditorForm, imgEditorForm };
